Add custom WP-CLI logger.
Most useful output such as "that plugin couldn't be activated" is
output as a warning. `--quiet` suppresses these, and piping errors to
`/dev/null` does, too.

Instead, we'll use a variant of WP-CLI's "quiet" logger - which
normally only prints errors - and get warnings printed as well. We can
then use the paramters on `exec` to check the responses for errors, and
handle throwing an exception in the driver appropriately.